Rename import dir to directive
